Prototype integration with JNet and its usage in ServletFactoryBean. Still lots of hacks and dirty code.
The most ugly hack here is URLStreamFactoryInstaller Spring bean that installs JNet
in its init method. ServletFactoryBean has dependency on this bean only to ensure that
init method is being called early enough.

Another problem is with usage of JNet as there is no way to release underlying Source instance.