Reworked HttpServletResponeBufferingWrapper a little bit.
The main idea behind all my changes is to move decision about buffering or not directly to our own implementation of ServletOutputStream.
ForwardingOrLimitingServletOutputStream acts in two modes depending on setting if it should buffer or just forward.
This change was needed because there might be following execution of methods for which previous implementation would fail:
1. stream = response.getOutputStream() //non buffering output stream is returned because by default status code is set to 200
2. response.setStatusCode(404)
3. stream.write() //write details about error, this is going to be written to response because non-buffering output stream is in use
4. response.resetBufferedResponse() //this fails with IllegalStateException

This problem was recorded in COCOON-2337 and now is fixed.

Also, test-case covering this problem was added and implementation of existing ones has been changed a little bit to more carefully check if buffering really works as expected.